requirements.txt
defusedxml==0.4.1
Django==1.10.5
django-allauth==0.30.0
django-rest-auth==0.9.0
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
mysqlclient==1.3.9
oauthlib==2.0.1
PyJWT==1.4.2
python3-openid==3.0.10
requests==2.12.5
requests-oauthlib==0.7.0
six==1.10.0
